+++ .github/workflows/go.yml
@@ -11,7 +11,7 @@ on:
 jobs:
 
   build:
-    name: Build and Unit Tests
+    name: Build and run Tests
     runs-on: ubuntu-latest
     steps:
     - name: Set up Go 1.16
@@ -21,17 +21,5 @@ jobs:
     - name: Check out code
       uses: actions/checkout@v1
     - name: Build
-      run: make unit-tests
-      
-  integration-tests:
-    name: Integration Tests
-    runs-on: ubuntu-latest
-    steps:
-    - name: Set up Go 1.16
-      uses: actions/setup-go@v2.1.3
-      with:
-        go-version: 1.16
-    - name: Check out code
-      uses: actions/checkout@v1
-    - name: Integration Tests
-      run: make integration-tests
+      run: make test
+
