+++ Makefile
@@ -1,13 +1,10 @@
-.PHONY: unit-tests integration-tests
+.PHONY: test
 
 override GOOS:=$(shell uname)
 override GO111MODULE=on
 
-unit-tests:
+test:
 	@CGO_ENABLED=0 go build && go test ./...
-
-integration-tests:
-	@go test -tags=integration ./...
 
 docker:
 	@docker build -t expediagroup/mittens:latest .
